fix configs
